+++ src/OpenTelemetry.Instrumentation.Http/Implementation/HttpHandlerDiagnosticListener.cs
@@ -29,6 +29,10 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
 {
     internal class HttpHandlerDiagnosticListener : ListenerHandler
     {
+        public const string RequestCustomPropertyName = "OTel.HttpHandler.Request";
+        public const string ResponseCustomPropertyName = "OTel.HttpHandler.Response";
+        public const string ExceptionCustomPropertyName = "OTel.HttpHandler.Exception";
+
         private static readonly Func<HttpRequestMessage, string, IEnumerable<string>> HttpRequestMessageHeaderValuesGetter = (request, name) =>
         {
             if (request.Headers.TryGetValues(name, out var values))
@@ -81,7 +85,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
                 return;
             }
 
-            if (this.options.TextFormat.IsInjected(request, HttpRequestMessageHeaderValuesGetter))
+            if (this.options.TextFormat.Extract(default, request, HttpRequestMessageHeaderValuesGetter) != default)
             {
                 // this request is already instrumented, we should back off
                 activity.IsAllDataRequested = false;
@@ -90,6 +94,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
 
             activity.SetKind(ActivityKind.Client);
             activity.DisplayName = HttpTagHelper.GetOperationNameForHttpMethod(request.Method);
+            activity.SetCustomProperty(RequestCustomPropertyName, request);
 
             this.activitySource.Start(activity);
 
@@ -107,7 +112,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
 
             if (!(this.httpClientSupportsW3C && this.options.TextFormat is TraceContextFormat))
             {
-                this.options.TextFormat.Inject(activity.Context, request, HttpRequestMessageHeaderValueSetter);
+                this.options.TextFormat.Inject(new PropagationContext(activity.Context, activity.Baggage), request, HttpRequestMessageHeaderValueSetter);
             }
         }
 
@@ -135,7 +140,8 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
 
                 if (this.stopResponseFetcher.Fetch(payload) is HttpResponseMessage response)
                 {
-                    // response could be null for DNS issues, timeouts, etc...
+                    activity.SetCustomProperty(ResponseCustomPropertyName, response);
+
                     activity.SetTag(SemanticConventions.AttributeHttpStatusCode, (int)response.StatusCode);
 
                     activity.SetStatus(
@@ -157,6 +163,8 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
                     HttpInstrumentationEventSource.Log.NullPayload(nameof(HttpHandlerDiagnosticListener), nameof(this.OnException));
                     return;
                 }
+
+                activity.SetCustomProperty(ExceptionCustomPropertyName, exc);
 
                 if (exc is HttpRequestException)
                 {